@@ -1,21 +1,22 @@
 name: simple_live_core
-version: 1.0.2
+version: 1.0.3
 description: "聚合直播核心库"
-repository: https://github.com/xiaoyaocz/simple_live_client
+repository: https://github.com/xiaoyaocz/dart_simple_live
 publish_to: "none"
 environment: 
-  sdk: '>=2.19.1 <3.0.0'
+  sdk: '>=3.0.5 <4.0.0'
 
 dependencies: 
-  dio: ^5.3.2
-  logger: ^2.0.2
-  web_socket_channel: ^2.4.0
+  dio: ^5.5.0+1
+  logger: ^2.4.0
+  web_socket_channel: ^3.0.1
   html_unescape: ^2.0.0
   protobuf: ^3.1.0
   crypto: ^3.0.3
   brotli: ^0.6.0
-  dart_tars_protocol:
-    git: https://github.com/xiaoyaocz/dart_tars_protocol.git
+  tars_dart:
+    path: ./packages/tars_dart
+  fixnum: ^1.1.0
 
 dev_dependencies: 
   lints: ^2.0.0
