@@ -1,9 +1,30 @@
+import 'dart:io';
+
+import 'package:device_info_plus/device_info_plus.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
+import 'package:intl/intl.dart';
 import 'package:logger/logger.dart';
+import 'package:path_provider/path_provider.dart';
+import 'package:simple_live_app/app/utils.dart';
 
 class Log {
+  static LogFileWriter? logFileWriter;
+  static void initWriter() {
+    logFileWriter = LogFileWriter();
+  }
+
+  static void disposeWriter() {
+    logFileWriter?.close();
+    logFileWriter = null;
+  }
+
+  static void writeLog(content, [Level level = Level.info]) {
+    logFileWriter
+        ?.write("[${level.name.toUpperCase()}] $_currentTime：$content");
+  }
+
   static RxList<DebugLogModel> debugLogs = <DebugLogModel>[].obs;
 
   static void addDebugLog(String content, Color? color) {
@@ -33,32 +54,103 @@ class Log {
     ),
   );
 
-  static void d(String message) {
+  static void d(String message, [bool writeFile = true]) {
     addDebugLog(message, Colors.orange);
     logger.d("${DateTime.now().toString()}\n$message");
+    if (writeFile) {
+      writeLog(message, Level.debug);
+    }
   }
 
-  static void i(String message) {
+  static void i(String message, [bool writeFile = true]) {
     addDebugLog(message, Colors.blue);
     logger.i("${DateTime.now().toString()}\n$message");
+    if (writeFile) {
+      logFileWriter?.write("[INFO] $_currentTime：$message");
+      writeLog(message, Level.info);
+    }
   }
 
-  static void e(String message, StackTrace stackTrace) {
+  static void e(String message, StackTrace stackTrace,
+      [bool writeFile = true]) {
     addDebugLog('$message\r\n\r\n$stackTrace', Colors.red);
     logger.e("${DateTime.now().toString()}\n$message", stackTrace: stackTrace);
+    if (writeFile) {
+      writeLog("$message\n$stackTrace", Level.error);
+    }
   }
 
-  static void w(String message) {
+  static void w(String message, [bool writeFile = true]) {
     addDebugLog(message, Colors.pink);
     logger.w("${DateTime.now().toString()}\n$message");
+    if (writeFile) {
+      writeLog(message, Level.warning);
+    }
   }
 
-  static void logPrint(dynamic obj) {
+  static void logPrint(dynamic obj, [bool writeFile = true]) {
     addDebugLog(obj.toString(), Colors.red);
+    if (writeFile) {
+      writeLog(obj, Level.info);
+    }
     //logger.e(obj.toString(), obj, obj?.stackTrace);
     if (kDebugMode) {
       print(obj);
     }
+  }
+
+  static String get _currentTime => Utils.timeFormat.format(DateTime.now());
+}
+
+class LogFileWriter {
+  late String fileName;
+  LogFileWriter() {
+    var dt = DateFormat("yyyy-MM-dd HH-mm-ss").format(DateTime.now());
+    fileName = "$dt.log";
+    initFile();
+  }
+  IOSink? fileWriter;
+  void initFile() async {
+    var supportDir = await getApplicationSupportDirectory();
+    var logDir = Directory("${supportDir.path}/log");
+    if (!await logDir.exists()) {
+      await logDir.create();
+    }
+    var logFile = File("${logDir.path}/$fileName");
+    fileWriter = logFile.openWrite(mode: FileMode.append);
+    writeSystemInfo();
+  }
+
+  void write(String content) {
+    fileWriter?.write(content);
+    fileWriter?.write("\r\n");
+  }
+
+  Future close() async {
+    await fileWriter?.close();
+  }
+
+  void writeSystemInfo() async {
+    DeviceInfoPlugin deviceInfo = DeviceInfoPlugin();
+    write("System Info:");
+    write("Current Time: ${DateTime.now()}");
+    write("Platform: ${Platform.operatingSystem}");
+    write("Version: ${Platform.operatingSystemVersion}");
+    write("Local: ${Platform.localeName}");
+    write(
+        "App Version: ${Utils.packageInfo.version}+${Utils.packageInfo.buildNumber}");
+    if (Platform.isAndroid) {
+      write((await deviceInfo.androidInfo).data.toString());
+    } else if (Platform.isIOS) {
+      write((await deviceInfo.iosInfo).data.toString());
+    } else if (Platform.isLinux) {
+      write((await deviceInfo.linuxInfo).data.toString());
+    } else if (Platform.isMacOS) {
+      write((await deviceInfo.macOsInfo).data.toString());
+    } else if (Platform.isWindows) {
+      write((await deviceInfo.windowsInfo).data.toString());
+    }
+    write("End System Info");
   }
 }
 
