@@ -15,10 +15,13 @@ import 'package:simple_live_app/app/log.dart';
 import 'package:simple_live_app/requests/common_request.dart';
 import 'package:url_launcher/url_launcher_string.dart';
 
+typedef TextValidate = bool Function(String text);
+
 class Utils {
   static late PackageInfo packageInfo;
   static DateFormat dateFormat = DateFormat("MM-dd HH:mm");
   static DateFormat dateFormatWithYear = DateFormat("yyyy-MM-dd HH:mm");
+  static DateFormat timeFormat = DateFormat("HH:mm:ss");
 
   /// 处理时间
   static String parseTime(DateTime? dt) {
@@ -181,12 +184,12 @@ class Utils {
     SmartDialog.dismiss(status: SmartStatus.allCustom);
   }
 
-  static void showBottomSheet({
+  static Future showBottomSheet({
     required String title,
     required Widget child,
     double maxWidth = 600,
-  }) {
-    showModalBottomSheet(
+  }) async {
+    var result = await showModalBottomSheet(
       context: Get.context!,
       constraints: BoxConstraints(
         maxWidth: maxWidth,
@@ -215,6 +218,7 @@ class Utils {
         ],
       ),
     );
+    return result;
   }
 
   /// 文本编辑的弹窗
@@ -222,11 +226,14 @@ class Utils {
   /// - `title` 弹窗标题
   /// - `confirm` 确认按钮内容
   /// - `cancel` 取消按钮内容
-  static Future<String?> showEditTextDialog(String content,
-      {String title = '',
-      String? hintText,
-      String confirm = '',
-      String cancel = ''}) async {
+  static Future<String?> showEditTextDialog(
+    String content, {
+    String title = '',
+    String? hintText,
+    String confirm = '',
+    String cancel = '',
+    TextValidate? validate,
+  }) async {
     final TextEditingController textEditingController =
         TextEditingController(text: content);
     var result = await Get.dialog(
@@ -255,6 +262,10 @@ class Utils {
           ),
           TextButton(
             onPressed: () {
+              if (validate != null && !validate(textEditingController.text)) {
+                return;
+              }
+
               Get.back(result: textEditingController.text);
             },
             child: const Text("确定"),
@@ -517,5 +528,28 @@ class Utils {
       SmartDialog.showToast("读取剪切板内容失败：$e");
     }
     return null;
+  }
+
+  static bool isRegexFormat(String keyword) {
+    return keyword.startsWith('/') &&
+        keyword.endsWith('/') &&
+        keyword.length > 2;
+  }
+
+  static String removeRegexFormat(String keyword) {
+    return keyword.substring(1, keyword.length - 1);
+  }
+
+  static String parseFileSize(int size) {
+    if (size < 1024) {
+      return "$size B";
+    }
+    if (size < 1024 * 1024) {
+      return "${(size / 1024).toStringAsFixed(2)} KB";
+    }
+    if (size < 1024 * 1024 * 1024) {
+      return "${(size / 1024 / 1024).toStringAsFixed(2)} MB";
+    }
+    return "${(size / 1024 / 1024 / 1024).toStringAsFixed(2)} GB";
   }
 }
