@@ -9,9 +9,9 @@ import 'package:simple_live_app/modules/category/category_controller.dart';
 import 'package:simple_live_app/modules/category/category_page.dart';
 import 'package:simple_live_app/modules/home/home_controller.dart';
 import 'package:simple_live_app/modules/home/home_page.dart';
-import 'package:simple_live_app/modules/user/follow_user/follow_user_controller.dart';
-import 'package:simple_live_app/modules/user/follow_user/follow_user_page.dart';
-import 'package:simple_live_app/modules/user/user_page.dart';
+import 'package:simple_live_app/modules/follow_user/follow_user_controller.dart';
+import 'package:simple_live_app/modules/follow_user/follow_user_page.dart';
+import 'package:simple_live_app/modules/mine/mine_page.dart';
 
 class IndexedController extends GetxController {
   RxList<HomePageItem> items = RxList<HomePageItem>([]);
@@ -40,7 +40,7 @@ class IndexedController extends GetxController {
           pages[i] = const CategoryPage();
           break;
         case 3:
-          pages[i] = const UserPage();
+          pages[i] = const MinePage();
           break;
         default:
       }
