@@ -1,6 +1,7 @@
 import 'dart:io';
-
+import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:get/get.dart';
 import 'package:media_kit_video/media_kit_video.dart';
 import 'package:ns_danmaku/ns_danmaku.dart';
@@ -10,8 +11,11 @@ import 'package:simple_live_app/app/controller/app_settings_controller.dart';
 import 'package:simple_live_app/app/sites.dart';
 import 'package:simple_live_app/app/utils.dart';
 import 'package:simple_live_app/modules/live_room/live_room_controller.dart';
-import 'package:simple_live_app/modules/user/danmu_settings_page.dart';
+import 'package:simple_live_app/modules/settings/danmu_settings_page.dart';
+import 'package:simple_live_app/services/follow_service.dart';
+import 'package:simple_live_app/widgets/desktop_refresh_button.dart';
 import 'package:simple_live_app/widgets/follow_user_item.dart';
+import 'package:window_manager/window_manager.dart';
 
 Widget playerControls(
   VideoState videoState,
@@ -37,288 +41,334 @@ Widget buildFullControls(
   LiveRoomController controller,
 ) {
   var padding = MediaQuery.of(videoState.context).padding;
+  GlobalKey volumeButtonkey = GlobalKey();
+  return DragToMoveArea(
+    child: Stack(
+      children: [
+        Container(),
+        buildDanmuView(videoState, controller),
 
-  return Stack(
-    children: [
-      Container(),
-      buildDanmuView(videoState, controller),
-
-      Center(
-        child: // 中间
-            StreamBuilder(
-          stream: videoState.widget.controller.player.stream.buffering,
-          initialData: videoState.widget.controller.player.state.buffering,
-          builder: (_, s) => Visibility(
-            visible: s.data ?? false,
-            child: const Center(
-              child: CircularProgressIndicator(),
-            ),
-          ),
-        ),
-      ),
-      Positioned.fill(
-        child: GestureDetector(
-          onTap: controller.onTap,
-          onDoubleTapDown: controller.onDoubleTap,
-          onLongPress: () {
-            showFollowUser(controller);
-          },
-          onVerticalDragStart: controller.onVerticalDragStart,
-          onVerticalDragUpdate: controller.onVerticalDragUpdate,
-          onVerticalDragEnd: controller.onVerticalDragEnd,
-          child: Container(
-            width: double.infinity,
-            height: double.infinity,
-            color: Colors.transparent,
-          ),
-        ),
-      ),
-
-      // 顶部
-      Obx(
-        () => AnimatedPositioned(
-          left: 0,
-          right: 0,
-          top: (controller.showControlsState.value &&
-                  !controller.lockControlsState.value)
-              ? 0
-              : -(48 + padding.top),
-          duration: const Duration(milliseconds: 200),
-          child: Container(
-            height: 48 + padding.top,
-            padding: EdgeInsets.only(
-              left: padding.left + 12,
-              right: padding.right + 12,
-              top: padding.top,
-            ),
-            decoration: const BoxDecoration(
-              gradient: LinearGradient(
-                begin: Alignment.bottomCenter,
-                end: Alignment.topCenter,
-                colors: [
-                  Colors.transparent,
-                  Colors.black87,
-                ],
+        Center(
+          child: // 中间
+              StreamBuilder(
+            stream: videoState.widget.controller.player.stream.buffering,
+            initialData: videoState.widget.controller.player.state.buffering,
+            builder: (_, s) => Visibility(
+              visible: s.data ?? false,
+              child: const Center(
+                child: CircularProgressIndicator(),
               ),
             ),
-            child: Row(
-              children: [
-                IconButton(
-                  onPressed: controller.exitFull,
-                  icon: const Icon(
-                    Icons.arrow_back,
-                    color: Colors.white,
-                    size: 24,
-                  ),
+          ),
+        ),
+        Positioned.fill(
+          child: GestureDetector(
+            onTap: controller.onTap,
+            onDoubleTapDown: controller.onDoubleTap,
+            onLongPress: () {
+              if (controller.lockControlsState.value) {
+                return;
+              }
+              showFollowUser(controller);
+            },
+            onVerticalDragStart: controller.onVerticalDragStart,
+            onVerticalDragUpdate: controller.onVerticalDragUpdate,
+            onVerticalDragEnd: controller.onVerticalDragEnd,
+            child: MouseRegion(
+              onHover: (PointerHoverEvent event) {
+                controller.onHover(event, videoState.context);
+              },
+              child: Container(
+                width: double.infinity,
+                height: double.infinity,
+                color: Colors.transparent,
+                // child: Visibility(
+                //   //拖拽区域
+                //   visible: controller.smallWindowState.value,
+                //   child: DragToMoveArea(
+                //       child: Container(
+                //     width: double.infinity,
+                //     height: double.infinity,
+                //     color: Colors.transparent,
+                //   )),
+                // ),
+              ),
+            ),
+          ),
+        ),
+
+        // 顶部
+        Obx(
+          () => AnimatedPositioned(
+            left: 0,
+            right: 0,
+            top: (controller.showControlsState.value &&
+                    !controller.lockControlsState.value)
+                ? 0
+                : -(48 + padding.top),
+            duration: const Duration(milliseconds: 200),
+            child: Container(
+              height: 48 + padding.top,
+              padding: EdgeInsets.only(
+                left: padding.left + 12,
+                right: padding.right + 12,
+                top: padding.top,
+              ),
+              decoration: const BoxDecoration(
+                gradient: LinearGradient(
+                  begin: Alignment.bottomCenter,
+                  end: Alignment.topCenter,
+                  colors: [
+                    Colors.transparent,
+                    Colors.black87,
+                  ],
                 ),
-                AppStyle.hGap12,
-                Expanded(
-                  child: Text(
-                    "${controller.detail.value?.title} - ${controller.detail.value?.userName}",
-                    maxLines: 1,
-                    overflow: TextOverflow.ellipsis,
-                    style: const TextStyle(color: Colors.white, fontSize: 16),
-                  ),
-                ),
-                AppStyle.hGap12,
-                IconButton(
-                  onPressed: () {
-                    controller.saveScreenshot();
-                  },
-                  icon: const Icon(
-                    Icons.camera_alt_outlined,
-                    color: Colors.white,
-                    size: 24,
-                  ),
-                ),
-                IconButton(
-                  onPressed: () {
-                    showFollowUser(controller);
-                  },
-                  icon: const Icon(
-                    Remix.play_list_2_line,
-                    color: Colors.white,
-                    size: 24,
-                  ),
-                ),
-                Visibility(
-                  visible: Platform.isAndroid,
-                  child: IconButton(
+              ),
+              child: Row(
+                children: [
+                  IconButton(
                     onPressed: () {
-                      controller.enablePIP();
+                      if (controller.smallWindowState.value) {
+                        controller.exitSmallWindow();
+                      } else {
+                        controller.exitFull();
+                      }
                     },
                     icon: const Icon(
-                      Icons.picture_in_picture,
+                      Icons.arrow_back,
                       color: Colors.white,
                       size: 24,
                     ),
                   ),
-                ),
-                IconButton(
-                  onPressed: () {
-                    showPlayerSettings(controller);
-                  },
-                  icon: const Icon(
-                    Icons.more_horiz,
-                    color: Colors.white,
-                    size: 24,
+                  AppStyle.hGap12,
+                  Expanded(
+                    child: Text(
+                      "${controller.detail.value?.title} - ${controller.detail.value?.userName}",
+                      maxLines: 1,
+                      overflow: TextOverflow.ellipsis,
+                      style: const TextStyle(color: Colors.white, fontSize: 16),
+                    ),
                   ),
-                ),
-              ],
-            ),
-          ),
-        ),
-      ),
-      // 底部
-      Obx(
-        () => AnimatedPositioned(
-          left: 0,
-          right: 0,
-          bottom: (controller.showControlsState.value &&
-                  !controller.lockControlsState.value)
-              ? 0
-              : -(80 + padding.bottom),
-          duration: const Duration(milliseconds: 200),
-          child: Container(
-            decoration: const BoxDecoration(
-              gradient: LinearGradient(
-                begin: Alignment.topCenter,
-                end: Alignment.bottomCenter,
-                colors: [
-                  Colors.transparent,
-                  Colors.black87,
+                  AppStyle.hGap12,
+                  IconButton(
+                    onPressed: () {
+                      controller.saveScreenshot();
+                    },
+                    icon: const Icon(
+                      Icons.camera_alt_outlined,
+                      color: Colors.white,
+                      size: 24,
+                    ),
+                  ),
+                  IconButton(
+                    onPressed: () {
+                      showFollowUser(controller);
+                    },
+                    icon: const Icon(
+                      Remix.play_list_2_line,
+                      color: Colors.white,
+                      size: 24,
+                    ),
+                  ),
+                  Visibility(
+                    visible: Platform.isAndroid,
+                    child: IconButton(
+                      onPressed: () {
+                        controller.enablePIP();
+                      },
+                      icon: const Icon(
+                        Icons.picture_in_picture,
+                        color: Colors.white,
+                        size: 24,
+                      ),
+                    ),
+                  ),
+                  IconButton(
+                    onPressed: () {
+                      showPlayerSettings(controller);
+                    },
+                    icon: const Icon(
+                      Icons.more_horiz,
+                      color: Colors.white,
+                      size: 24,
+                    ),
+                  ),
                 ],
               ),
             ),
-            padding: EdgeInsets.only(
-              left: padding.left + 12,
-              right: padding.right + 12,
-              bottom: padding.bottom,
-            ),
-            child: Row(
-              children: [
-                IconButton(
-                  onPressed: () {
-                    controller.refreshRoom();
-                  },
-                  icon: const Icon(
-                    Remix.refresh_line,
-                    color: Colors.white,
-                  ),
+          ),
+        ),
+        // 底部
+        Obx(
+          () => AnimatedPositioned(
+            left: 0,
+            right: 0,
+            bottom: (controller.showControlsState.value &&
+                    !controller.lockControlsState.value)
+                ? 0
+                : -(80 + padding.bottom),
+            duration: const Duration(milliseconds: 200),
+            child: Container(
+              decoration: const BoxDecoration(
+                gradient: LinearGradient(
+                  begin: Alignment.topCenter,
+                  end: Alignment.bottomCenter,
+                  colors: [
+                    Colors.transparent,
+                    Colors.black87,
+                  ],
                 ),
-                Offstage(
-                  offstage: controller.showDanmakuState.value,
-                  child: IconButton(
-                    onPressed: () => controller.showDanmakuState.value =
-                        !controller.showDanmakuState.value,
+              ),
+              padding: EdgeInsets.only(
+                left: padding.left + 12,
+                right: padding.right + 12,
+                bottom: padding.bottom,
+              ),
+              child: Row(
+                children: [
+                  IconButton(
+                    onPressed: () {
+                      controller.refreshRoom();
+                    },
+                    icon: const Icon(
+                      Remix.refresh_line,
+                      color: Colors.white,
+                    ),
+                  ),
+                  Offstage(
+                    offstage: controller.showDanmakuState.value,
+                    child: IconButton(
+                      onPressed: () => controller.showDanmakuState.value =
+                          !controller.showDanmakuState.value,
+                      icon: const ImageIcon(
+                        AssetImage('assets/icons/icon_danmaku_open.png'),
+                        size: 24,
+                        color: Colors.white,
+                      ),
+                    ),
+                  ),
+                  Offstage(
+                    offstage: !controller.showDanmakuState.value,
+                    child: IconButton(
+                      onPressed: () => controller.showDanmakuState.value =
+                          !controller.showDanmakuState.value,
+                      icon: const ImageIcon(
+                        AssetImage('assets/icons/icon_danmaku_close.png'),
+                        size: 24,
+                        color: Colors.white,
+                      ),
+                    ),
+                  ),
+                  IconButton(
+                    onPressed: () {
+                      showDanmakuSettings(controller);
+                    },
                     icon: const ImageIcon(
-                      AssetImage('assets/icons/icon_danmaku_open.png'),
+                      AssetImage('assets/icons/icon_danmaku_setting.png'),
                       size: 24,
                       color: Colors.white,
                     ),
                   ),
-                ),
-                Offstage(
-                  offstage: !controller.showDanmakuState.value,
-                  child: IconButton(
-                    onPressed: () => controller.showDanmakuState.value =
-                        !controller.showDanmakuState.value,
-                    icon: const ImageIcon(
-                      AssetImage('assets/icons/icon_danmaku_close.png'),
-                      size: 24,
-                      color: Colors.white,
+                  const Expanded(child: Center()),
+                  Visibility(
+                    visible: !Platform.isAndroid && !Platform.isIOS,
+                    child: IconButton(
+                      key: volumeButtonkey,
+                      onPressed: () {
+                        controller
+                            .showVolumeSlider(volumeButtonkey.currentContext!);
+                      },
+                      icon: const Icon(
+                        Icons.volume_down,
+                        size: 24,
+                        color: Colors.white,
+                      ),
                     ),
                   ),
-                ),
-                IconButton(
-                  onPressed: () {
-                    showDanmakuSettings(controller);
-                  },
-                  icon: const ImageIcon(
-                    AssetImage('assets/icons/icon_danmaku_setting.png'),
-                    size: 24,
-                    color: Colors.white,
+                  TextButton(
+                    onPressed: () {
+                      showQualitesInfo(controller);
+                    },
+                    child: Obx(
+                      () => Text(
+                        controller.currentQualityInfo.value,
+                        style:
+                            const TextStyle(color: Colors.white, fontSize: 15),
+                      ),
+                    ),
                   ),
-                ),
-                const Expanded(child: Center()),
-                TextButton(
-                  onPressed: () {
-                    showQualitesInfo(controller);
-                  },
-                  child: Obx(
-                    () => Text(
-                      controller.currentQualityInfo.value,
+                  TextButton(
+                    onPressed: () {
+                      showLinesInfo(controller);
+                    },
+                    child: Text(
+                      controller.currentLineInfo.value,
                       style: const TextStyle(color: Colors.white, fontSize: 15),
                     ),
                   ),
-                ),
-                TextButton(
-                  onPressed: () {
-                    showLinesInfo(controller);
-                  },
-                  child: Text(
-                    controller.currentLineInfo.value,
-                    style: const TextStyle(color: Colors.white, fontSize: 15),
+                  IconButton(
+                    onPressed: () {
+                      if (controller.smallWindowState.value) {
+                        controller.exitSmallWindow();
+                      } else {
+                        controller.exitFull();
+                      }
+                    },
+                    icon: const Icon(
+                      Remix.fullscreen_exit_fill,
+                      color: Colors.white,
+                    ),
                   ),
-                ),
-                IconButton(
-                  onPressed: () {
-                    controller.exitFull();
-                  },
-                  icon: const Icon(
-                    Remix.fullscreen_exit_fill,
-                    color: Colors.white,
-                  ),
-                ),
-              ],
+                ],
+              ),
             ),
           ),
         ),
-      ),
 
-      // 右侧锁定
-      Obx(
-        () => AnimatedPositioned(
-          top: 0,
-          bottom: 0,
-          right: controller.showControlsState.value
-              ? padding.right + 12
-              : -(64 + padding.right),
-          duration: const Duration(milliseconds: 200),
-          child: buildLockButton(controller),
+        // 右侧锁定
+        Obx(
+          () => AnimatedPositioned(
+            top: 0,
+            bottom: 0,
+            right: controller.showControlsState.value
+                ? padding.right + 12
+                : -(64 + padding.right),
+            duration: const Duration(milliseconds: 200),
+            child: buildLockButton(controller),
+          ),
         ),
-      ),
-      // 左侧锁定
-      Obx(
-        () => AnimatedPositioned(
-          top: 0,
-          bottom: 0,
-          left: controller.showControlsState.value
-              ? padding.left + 12
-              : -(64 + padding.right),
-          duration: const Duration(milliseconds: 200),
-          child: buildLockButton(controller),
+        // 左侧锁定
+        Obx(
+          () => AnimatedPositioned(
+            top: 0,
+            bottom: 0,
+            left: controller.showControlsState.value
+                ? padding.left + 12
+                : -(64 + padding.right),
+            duration: const Duration(milliseconds: 200),
+            child: buildLockButton(controller),
+          ),
         ),
-      ),
-      Obx(
-        () => Offstage(
-          offstage: !controller.showGestureTip.value,
-          child: Center(
-            child: Container(
-              padding: const EdgeInsets.all(12),
-              decoration: BoxDecoration(
-                color: Colors.grey.shade900,
-                borderRadius: BorderRadius.circular(12),
-              ),
-              child: Text(
-                controller.gestureTipText.value,
-                style: const TextStyle(fontSize: 18, color: Colors.white),
+        Obx(
+          () => Offstage(
+            offstage: !controller.showGestureTip.value,
+            child: Center(
+              child: Container(
+                padding: const EdgeInsets.all(12),
+                decoration: BoxDecoration(
+                  color: Colors.grey.shade900,
+                  borderRadius: BorderRadius.circular(12),
+                ),
+                child: Text(
+                  controller.gestureTipText.value,
+                  style: const TextStyle(fontSize: 18, color: Colors.white),
+                ),
               ),
             ),
           ),
         ),
-      ),
-    ],
+      ],
+    ),
   );
 }
 
@@ -354,6 +404,7 @@ Widget buildControls(
   VideoState videoState,
   LiveRoomController controller,
 ) {
+  GlobalKey volumeButtonkey = GlobalKey();
   return Stack(
     children: [
       Container(),
@@ -379,10 +430,13 @@ Widget buildControls(
           onVerticalDragUpdate: controller.onVerticalDragUpdate,
           onVerticalDragEnd: controller.onVerticalDragEnd,
           //onLongPress: controller.showDebugInfo,
-          child: Container(
-            width: double.infinity,
-            height: double.infinity,
-            color: Colors.transparent,
+          child: MouseRegion(
+            onEnter: controller.onEnter,
+            child: Container(
+              width: double.infinity,
+              height: double.infinity,
+              color: Colors.transparent,
+            ),
           ),
         ),
       ),
@@ -449,6 +503,22 @@ Widget buildControls(
                   ),
                 ),
                 const Expanded(child: Center()),
+                Visibility(
+                  visible: !Platform.isAndroid && !Platform.isIOS,
+                  child: IconButton(
+                    key: volumeButtonkey,
+                    onPressed: () {
+                      controller.showVolumeSlider(
+                        volumeButtonkey.currentContext!,
+                      );
+                    },
+                    icon: const Icon(
+                      Icons.volume_down,
+                      size: 24,
+                      color: Colors.white,
+                    ),
+                  ),
+                ),
                 Offstage(
                   offstage: isPortrait,
                   child: TextButton(
@@ -473,6 +543,19 @@ Widget buildControls(
                     child: Text(
                       controller.currentLineInfo.value,
                       style: const TextStyle(color: Colors.white, fontSize: 15),
+                    ),
+                  ),
+                ),
+                Visibility(
+                  visible: !Platform.isAndroid && !Platform.isIOS,
+                  child: IconButton(
+                    onPressed: () {
+                      controller.enterSmallWindow();
+                    },
+                    icon: const Icon(
+                      Icons.picture_in_picture,
+                      color: Colors.white,
+                      size: 24,
                     ),
                   ),
                 ),
@@ -731,35 +814,49 @@ void showFollowUser(LiveRoomController controller) {
     controller.showFollowUserSheet();
     return;
   }
-  //只显示开播直播间
-  controller.followController.setFilterMode(1);
+
   Utils.showRightDialog(
     title: "关注列表",
     width: 400,
     useSystem: true,
     child: Obx(
-      () => RefreshIndicator(
-        onRefresh: controller.followController.refreshData,
-        child: ListView.builder(
-          itemCount: controller.followController.list.length,
-          itemBuilder: (_, i) {
-            var item = controller.followController.list[i];
-            return Obx(
-              () => FollowUserItem(
-                item: item,
-                playing: controller.rxSite.value.id == item.siteId &&
-                    controller.rxRoomId.value == item.roomId,
-                onTap: () {
-                  Utils.hideRightDialog();
-                  controller.resetRoom(
-                    Sites.allSites[item.siteId]!,
-                    item.roomId,
-                  );
-                },
+      () => Stack(
+        children: [
+          RefreshIndicator(
+            onRefresh: FollowService.instance.loadData,
+            child: ListView.builder(
+              itemCount: FollowService.instance.liveList.length,
+              itemBuilder: (_, i) {
+                var item = FollowService.instance.liveList[i];
+                return Obx(
+                  () => FollowUserItem(
+                    item: item,
+                    playing: controller.rxSite.value.id == item.siteId &&
+                        controller.rxRoomId.value == item.roomId,
+                    onTap: () {
+                      Utils.hideRightDialog();
+                      controller.resetRoom(
+                        Sites.allSites[item.siteId]!,
+                        item.roomId,
+                      );
+                    },
+                  ),
+                );
+              },
+            ),
+          ),
+          if (Platform.isLinux || Platform.isWindows || Platform.isMacOS)
+            Positioned(
+              right: 12,
+              bottom: 12,
+              child: Obx(
+                () => DesktopRefreshButton(
+                  refreshing: FollowService.instance.updating.value,
+                  onPressed: FollowService.instance.loadData,
+                ),
               ),
-            );
-          },
-        ),
+            ),
+        ],
       ),
     ),
   );
