@@ -1,5 +1,7 @@
 import 'package:dio/dio.dart';
+import 'package:flutter/foundation.dart';
 import 'package:simple_live_app/app/log.dart';
+import 'package:simple_live_core/simple_live_core.dart';
 
 class CustomLogInterceptor extends Interceptor {
   @override
@@ -13,7 +15,8 @@ class CustomLogInterceptor extends Interceptor {
   void onError(DioException err, ErrorInterceptorHandler handler) {
     var time =
         DateTime.now().millisecondsSinceEpoch - err.requestOptions.extra["ts"];
-    Log.e('''【HTTP请求错误-${err.type}】 耗时:${time}ms
+    if (!kReleaseMode) {
+      Log.e('''【HTTP请求错误-${err.type}】 耗时:${time}ms
 ${err.message}
 
 Request Method：${err.requestOptions.method}
@@ -24,6 +27,20 @@ Request Data：${err.requestOptions.data}
 Request Headers：${err.requestOptions.headers}
 Response Headers：${err.response?.headers.map}
 Response Data：${err.response?.data}''', err.stackTrace);
+    } else {
+      CoreLog.e('''[HTTP Error] [${err.type}] [Time:${time}ms]
+${err.message}
+
+Request Method：${err.requestOptions.method}
+Response Code：${err.response?.statusCode}
+Request URL：${err.requestOptions.uri}
+Request Query：${err.requestOptions.queryParameters}
+Request Data：${err.requestOptions.data}
+Request Headers：${_maskHeader(err.requestOptions.headers)}
+Response Headers：${err.response?.headers.map}
+Response Data：${err.response?.data}''', err.stackTrace);
+    }
+
     super.onError(err, handler);
   }
 
@@ -31,8 +48,9 @@ Response Data：${err.response?.data}''', err.stackTrace);
   void onResponse(Response response, ResponseInterceptorHandler handler) {
     var time = DateTime.now().millisecondsSinceEpoch -
         response.requestOptions.extra["ts"];
-    Log.i(
-      '''【HTTP请求响应】 耗时:${time}ms
+    if (!kReleaseMode) {
+      Log.i(
+        '''【HTTP请求响应】 耗时:${time}ms
 Request Method：${response.requestOptions.method}
 Request Code：${response.statusCode}
 Request URL：${response.requestOptions.uri}
@@ -41,7 +59,26 @@ Request Data：${response.requestOptions.data}
 Request Headers：${response.requestOptions.headers}
 Response Headers：${response.headers.map}
 Response Data：${response.data}''',
-    );
+      );
+    } else {
+      CoreLog.i(
+        "[HTTP Response] [time:${time}ms] [${response.statusCode}] ${response.requestOptions.uri}",
+      );
+    }
     super.onResponse(response, handler);
+  }
+
+  // Header脱敏
+  String _maskHeader(Map<String, dynamic> header) {
+    var result = <String, dynamic>{};
+    header.forEach((key, value) {
+      var k = key.toLowerCase();
+      if (k == "cookie" || k == "authorization") {
+        result[key] = "******";
+      } else {
+        result[key] = value;
+      }
+    });
+    return result.toString();
   }
 }
