@@ -59,6 +59,9 @@ class LocalStorageService extends GetxService {
   /// 弹幕开启
   static const String kDanmuEnable = "DanmuEnable";
 
+  /// 弹幕字重
+  static const String kDanmuFontWeight = "DanmuFontWeight";
+
   /// 硬件解码
   static const String kHardwareDecode = "HardwareDecode";
 
@@ -83,14 +86,27 @@ class LocalStorageService extends GetxService {
   /// 定时关闭时间（分钟）
   static const String kAutoExitDuration = "AutoExitDuration";
 
+  /// 房间内定时关闭时间（分钟）
+  /// 需要一个不同的 key，因为用户在房间内设置的倒计时和全局的可能不同。
+  static const String kRoomAutoExitDuration = "RoomAutoExitDuration";
+
   /// 播放器兼容模式
   static const String kPlayerCompatMode = "PlayerCompatMode";
 
   /// 播放器后台自动暂停
   static const String kPlayerAutoPause = "PlayerAutoPause";
 
+  /// 播放器缓冲区大小
+  static const String kPlayerBufferSize = "PlayerBufferSize";
+
+  /// 播放器强制使用HTTPS
+  static const String kPlayerForceHttps = "PlayerForceHttps";
+
   /// 自动全屏
   static const String kAutoFullScreen = "AutoFullScreen";
+
+  /// 播放器音量
+  static const String kPlayerVolume = "PlayerVolume";
 
   /// 小窗隐藏弹幕
   static const String kPIPHideDanmu = "PIPHideDanmu";
@@ -107,6 +123,27 @@ class LocalStorageService extends GetxService {
   /// 提示哔哩哔哩登录
   static const String kBilibiliLoginTip = "BilibiliLoginTip";
 
+  /// 日志记录
+  static const String kLogEnable = "LogEnable";
+
+  /// 开启自定义播放器视频输出
+  static const String kCustomPlayerOutput = "CustomPlayerOutput";
+
+  /// 视频输出驱动
+  static const String kVideoOutputDriver = "VideoOutputDriver";
+
+  /// 视频硬件解码器
+  static const String kVideoHardwareDecoder = "VideoHardwareDecoder";
+
+  /// 开启自动更新关注
+  static const String kAutoUpdateFollowEnable = "AutoUpdateFollowEnable";
+
+  /// 定时自动更新关注间隔（分钟）
+  static const String kUpdateFollowDuration = "AutoUpdateFollowDuration";
+
+  /// 开启多线程更新关注
+  static const String kUpdateFollowThreadCount = "UpdateFollowThreadCount";
+
   late Box settingsBox;
   late Box<String> shieldBox;
 
@@ -120,9 +157,14 @@ class LocalStorageService extends GetxService {
   }
 
   T getValue<T>(dynamic key, T defaultValue) {
-    var value = settingsBox.get(key, defaultValue: defaultValue) as T;
-    Log.d("Get LocalStorage：$key\r\n$value");
-    return value;
+    try {
+      var value = settingsBox.get(key, defaultValue: defaultValue) as T;
+      Log.d("Get LocalStorage：$key\r\n$value");
+      return value;
+    } catch (e) {
+      Log.logPrint(e);
+      return defaultValue;
+    }
   }
 
   Future setValue<T>(dynamic key, T value) async {
