@@ -16,7 +16,7 @@ class DouyinSite implements LiveSite {
   LiveDanmaku getDanmaku() => DouyinDanmaku();
 
   static const String kDefaultUserAgent =
-      "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36 Edg/120.0.0.0";
+      "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/125.0.0.0 Safari/537.36 Edg/125.0.0.0";
 
   static const String kDefaultReferer = "https://live.douyin.com";
 
@@ -52,13 +52,13 @@ class DouyinSite implements LiveSite {
   Future<List<LiveCategory>> getCategores() async {
     List<LiveCategory> categories = [];
     var result = await HttpClient.instance.getText(
-      "https://live.douyin.com/hot_live",
+      "https://live.douyin.com/",
       queryParameters: {},
       header: await getRequestHeaders(),
     );
 
     var renderData =
-        RegExp(r'\{\\"pathname\\":\\"\/hot_live\\",\\"categoryData.*?\]\\n')
+        RegExp(r'\{\\"pathname\\":\\"\/\\",\\"categoryData.*?\]\\n')
                 .firstMatch(result)
                 ?.group(0) ??
             "";
@@ -174,64 +174,198 @@ class DouyinSite implements LiveSite {
 
   @override
   Future<LiveRoomDetail> getRoomDetail({required String roomId}) async {
-    var detail = await getRoomWebDetail(roomId);
-    var requestHeader = await getRequestHeaders();
-    var webRid = roomId;
-    var realRoomId =
-        detail["roomStore"]["roomInfo"]["room"]["id_str"].toString();
-    var userUniqueId = detail["userStore"]["odin"]["user_unique_id"].toString();
-    var result = await HttpClient.instance.getJson(
-      "https://live.douyin.com/webcast/room/web/enter/",
-      queryParameters: {
-        "aid": 6383,
-        "app_name": "douyin_web",
-        "live_id": 1,
-        "device_platform": "web",
-        "enter_from": "web_live",
-        "web_rid": webRid,
-        "room_id_str": realRoomId,
-        "enter_source": "",
-        "Room-Enter-User-Login-Ab": 0,
-        "is_need_double_stream": false,
-        "cookie_enabled": true,
-        "screen_width": 1980,
-        "screen_height": 1080,
-        "browser_language": "zh-CN",
-        "browser_platform": "Win32",
-        "browser_name": "Edge",
-        "browser_version": "120.0.0.0"
-      },
-      header: requestHeader,
-    );
-    var roomInfo = result["data"]["data"][0];
-    var userInfo = result["data"]["user"];
-    var roomStatus = (asT<int?>(roomInfo["status"]) ?? 0) == 2;
+    // 有两种roomId，一种是webRid，一种是roomId
+    // roomId是一次性的，用户每次重新开播都会生成一个新的roomId
+    // roomId一般长度为19位，例如：7376429659866598196
+    // webRid是固定的，用户每次开播都是同一个webRid
+    // webRid一般长度为11-12位，例如：416144012050
+    // 这里简单进行判断，如果roomId长度小于15，则认为是webRid
+    if (roomId.length <= 16) {
+      var webRid = roomId;
+      return await getRoomDetailByWebRid(webRid);
+    }
+
+    return await getRoomDetailByRoomId(roomId);
+  }
+
+  /// 通过roomId获取直播间信息
+  /// - [roomId] 直播间ID
+  /// - 返回直播间信息
+  Future<LiveRoomDetail> getRoomDetailByRoomId(String roomId) async {
+    // 读取房间信息
+    var roomData = await _getRoomDataByRoomId(roomId);
+
+    // 通过房间信息获取WebRid
+    var webRid = roomData["data"]["room"]["owner"]["web_rid"].toString();
+
+    // 读取用户唯一ID，用于弹幕连接
+    // 似乎这个参数不是必须的，先随机生成一个
+    //var userUniqueId = await _getUserUniqueId(webRid);
+    var userUniqueId = generateRandomNumber(12).toString();
+
+    var room = roomData["data"]["room"];
+    var owner = room["owner"];
+
+    var status = asT<int?>(room["status"]) ?? 0;
+
+    // roomId是一次性的，用户每次重新开播都会生成一个新的roomId
+    // 所以如果roomId对应的直播间状态不是直播中，就通过webRid获取直播间信息
+    if (status == 4) {
+      var result = await getRoomDetailByWebRid(webRid);
+      return result;
+    }
+
+    var roomStatus = status == 2;
+    // 主要是为了获取cookie,用于弹幕websocket连接
+    var headers = await getRequestHeaders();
+
     return LiveRoomDetail(
-      roomId: roomId,
-      title: roomInfo["title"].toString(),
-      cover: roomStatus ? roomInfo["cover"]["url_list"][0].toString() : "",
-      userName: userInfo["nickname"].toString(),
-      userAvatar: userInfo["avatar_thumb"]["url_list"][0].toString(),
+      roomId: webRid,
+      title: room["title"].toString(),
+      cover: roomStatus ? room["cover"]["url_list"][0].toString() : "",
+      userName: owner["nickname"].toString(),
+      userAvatar: owner["avatar_thumb"]["url_list"][0].toString(),
       online: roomStatus
-          ? asT<int?>(roomInfo["room_view_stats"]["display_value"]) ?? 0
+          ? asT<int?>(room["room_view_stats"]["display_value"]) ?? 0
           : 0,
       status: roomStatus,
       url: "https://live.douyin.com/$webRid",
-      introduction: roomInfo["title"].toString(),
+      introduction: owner["signature"].toString(),
       notice: "",
       danmakuData: DouyinDanmakuArgs(
         webRid: webRid,
-        roomId: realRoomId,
+        roomId: roomId,
         userId: userUniqueId,
         cookie: headers["cookie"],
       ),
-      data: roomInfo["stream_url"],
+      data: room["stream_url"],
     );
   }
 
-  Future<Map> getRoomWebDetail(String webRid) async {
-    var headResp = await HttpClient.instance
-        .head("https://live.douyin.com/$webRid", header: headers);
+  /// 通过WebRid获取直播间信息
+  /// - [webRid] 直播间RID
+  /// - 返回直播间信息
+  Future<LiveRoomDetail> getRoomDetailByWebRid(String webRid) async {
+    try {
+      var result = await _getRoomDetailByWebRidApi(webRid);
+      return result;
+    } catch (e) {
+      CoreLog.error(e);
+    }
+    return await _getRoomDetailByWebRidHtml(webRid);
+  }
+
+  /// 通过WebRid访问直播间API，从API中获取直播间信息
+  /// - [webRid] 直播间RID
+  /// - 返回直播间信息
+  Future<LiveRoomDetail> _getRoomDetailByWebRidApi(String webRid) async {
+    // 读取房间信息
+    var data = await _getRoomDataByApi(webRid);
+    var roomData = data["data"][0];
+    var userData = data["user"];
+    var roomId = roomData["id_str"].toString();
+
+    // 读取用户唯一ID，用于弹幕连接
+    // 似乎这个参数不是必须的，先随机生成一个
+    //var userUniqueId = await _getUserUniqueId(webRid);
+    var userUniqueId = generateRandomNumber(12).toString();
+
+    var owner = roomData["owner"];
+
+    var roomStatus = (asT<int?>(roomData["status"]) ?? 0) == 2;
+
+    // 主要是为了获取cookie,用于弹幕websocket连接
+    var headers = await getRequestHeaders();
+    return LiveRoomDetail(
+      roomId: webRid,
+      title: roomData["title"].toString(),
+      cover: roomStatus ? roomData["cover"]["url_list"][0].toString() : "",
+      userName: roomStatus
+          ? owner["nickname"].toString()
+          : userData["nickname"].toString(),
+      userAvatar: roomStatus
+          ? owner["avatar_thumb"]["url_list"][0].toString()
+          : userData["avatar_thumb"]["url_list"][0].toString(),
+      online: roomStatus
+          ? asT<int?>(roomData["room_view_stats"]["display_value"]) ?? 0
+          : 0,
+      status: roomStatus,
+      url: "https://live.douyin.com/$webRid",
+      introduction: owner?["signature"]?.toString() ?? "",
+      notice: "",
+      danmakuData: DouyinDanmakuArgs(
+        webRid: webRid,
+        roomId: roomId,
+        userId: userUniqueId,
+        cookie: headers["cookie"],
+      ),
+      data: roomStatus ? roomData["stream_url"] : {},
+    );
+  }
+
+  /// 通过WebRid访问直播间网页，从网页HTML中获取直播间信息
+  /// - [webRid] 直播间RID
+  /// - 返回直播间信息
+  Future<LiveRoomDetail> _getRoomDetailByWebRidHtml(String webRid) async {
+    var roomData = await _getRoomDataByHtml(webRid);
+    var roomId = roomData["roomStore"]["roomInfo"]["room"]["id_str"].toString();
+    var userUniqueId =
+        roomData["userStore"]["odin"]["user_unique_id"].toString();
+
+    var room = roomData["roomStore"]["roomInfo"]["room"];
+    var owner = room["owner"];
+    var anchor = roomData["roomStore"]["roomInfo"]["anchor"];
+    var roomStatus = (asT<int?>(room["status"]) ?? 0) == 2;
+
+    // 主要是为了获取cookie,用于弹幕websocket连接
+    var headers = await getRequestHeaders();
+
+    return LiveRoomDetail(
+      roomId: webRid,
+      title: room["title"].toString(),
+      cover: roomStatus ? room["cover"]["url_list"][0].toString() : "",
+      userName: roomStatus
+          ? owner["nickname"].toString()
+          : anchor["nickname"].toString(),
+      userAvatar: roomStatus
+          ? owner["avatar_thumb"]["url_list"][0].toString()
+          : anchor["avatar_thumb"]["url_list"][0].toString(),
+      online: roomStatus
+          ? asT<int?>(room["room_view_stats"]["display_value"]) ?? 0
+          : 0,
+      status: roomStatus,
+      url: "https://live.douyin.com/$webRid",
+      introduction: owner?["signature"]?.toString() ?? "",
+      notice: "",
+      danmakuData: DouyinDanmakuArgs(
+        webRid: webRid,
+        roomId: roomId,
+        userId: userUniqueId,
+        cookie: headers["cookie"],
+      ),
+      data: roomStatus ? room["stream_url"] : {},
+    );
+  }
+
+  /// 读取用户的唯一ID
+  /// - [webRid] 直播间RID
+  // ignore: unused_element
+  Future<String> _getUserUniqueId(String webRid) async {
+    try {
+      var webInfo = await _getRoomDataByHtml(webRid);
+      return webInfo["userStore"]["odin"]["user_unique_id"].toString();
+    } catch (e) {
+      return generateRandomNumber(12).toString();
+    }
+  }
+
+  /// 进入直播间前需要先获取cookie
+  /// - [webRid] 直播间RID
+  Future<String> _getWebCookie(String webRid) async {
+    var headResp = await HttpClient.instance.head(
+      "https://live.douyin.com/$webRid",
+      header: headers,
+    );
     var dyCookie = "";
     headResp.headers["set-cookie"]?.forEach((element) {
       var cookie = element.split(";")[0];
@@ -241,8 +375,17 @@ class DouyinSite implements LiveSite {
       if (cookie.contains("__ac_nonce")) {
         dyCookie += "$cookie;";
       }
+      if (cookie.contains("msToken")) {
+        dyCookie += "$cookie;";
+      }
     });
+    return dyCookie;
+  }
 
+  /// 通过webRid获取直播间Web信息
+  /// - [webRid] 直播间RID
+  Future<Map> _getRoomDataByHtml(String webRid) async {
+    var dyCookie = await _getWebCookie(webRid);
     var result = await HttpClient.instance.getText(
       "https://live.douyin.com/$webRid",
       queryParameters: {},
@@ -264,32 +407,124 @@ class DouyinSite implements LiveSite {
         .replaceAll(r"\\", r"\")
         .replaceAll(']\\n', "");
     var renderDataJson = json.decode(str);
-
     return renderDataJson["state"];
-    // return renderDataJson["app"]["initialState"]["roomStore"]["roomInfo"]
-    //         ["room"]["id_str"]
-    //     .toString();
+  }
+
+  /// 通过webRid获取直播间Web信息
+  /// - [webRid] 直播间RID
+  Future<Map> _getRoomDataByApi(String webRid) async {
+    var requestHeader = await getRequestHeaders();
+    var result = await HttpClient.instance.getJson(
+      "https://live.douyin.com/webcast/room/web/enter/",
+      queryParameters: {
+        "aid": 6383,
+        "app_name": "douyin_web",
+        "live_id": 1,
+        "device_platform": "web",
+        "enter_from": "web_live",
+        "web_rid": webRid,
+        "room_id_str": "",
+        "enter_source": "",
+        "Room-Enter-User-Login-Ab": 0,
+        "is_need_double_stream": false,
+        "cookie_enabled": true,
+        "screen_width": 1980,
+        "screen_height": 1080,
+        "browser_language": "zh-CN",
+        "browser_platform": "Win32",
+        "browser_name": "Edge",
+        "browser_version": "125.0.0.0"
+      },
+      header: requestHeader,
+    );
+
+    return result["data"];
+  }
+
+  /// 通过roomId获取直播间信息
+  /// - [roomId] 直播间ID
+  Future<Map> _getRoomDataByRoomId(String roomId) async {
+    var result = await HttpClient.instance.getJson(
+      'https://webcast.amemv.com/webcast/room/reflow/info/',
+      queryParameters: {
+        "type_id": 0,
+        "live_id": 1,
+        "room_id": roomId,
+        "sec_user_id": "",
+        "version_code": "99.99.99",
+        "app_id": 6383,
+      },
+      header: await getRequestHeaders(),
+    );
+    return result;
   }
 
   @override
   Future<List<LivePlayQuality>> getPlayQualites(
       {required LiveRoomDetail detail}) async {
     List<LivePlayQuality> qualities = [];
-    var qualityData = json.decode(
-        detail.data["live_core_sdk_data"]["pull_data"]["stream_data"])["data"];
+
     var qulityList =
         detail.data["live_core_sdk_data"]["pull_data"]["options"]["qualities"];
-    for (var quality in qulityList) {
-      var qualityItem = LivePlayQuality(
-        quality: quality["name"],
-        sort: quality["level"],
-        data: <String>[
-          qualityData[quality["sdk_key"]]["main"]["flv"].toString(),
-          qualityData[quality["sdk_key"]]["main"]["hls"].toString(),
-        ],
-      );
-      qualities.add(qualityItem);
+    var streamData = detail.data["live_core_sdk_data"]["pull_data"]
+            ["stream_data"]
+        .toString();
+
+    if (!streamData.startsWith('{')) {
+      var flvList =
+          (detail.data["flv_pull_url"] as Map).values.cast<String>().toList();
+      var hlsList = (detail.data["hls_pull_url_map"] as Map)
+          .values
+          .cast<String>()
+          .toList();
+      for (var quality in qulityList) {
+        int level = quality["level"];
+        List<String> urls = [];
+        var flvIndex = flvList.length - level;
+        if (flvIndex >= 0 && flvIndex < flvList.length) {
+          urls.add(flvList[flvIndex]);
+        }
+        var hlsIndex = hlsList.length - level;
+        if (hlsIndex >= 0 && hlsIndex < hlsList.length) {
+          urls.add(hlsList[hlsIndex]);
+        }
+        var qualityItem = LivePlayQuality(
+          quality: quality["name"],
+          sort: level,
+          data: urls,
+        );
+        if (urls.isNotEmpty) {
+          qualities.add(qualityItem);
+        }
+      }
+    } else {
+      var qualityData = json.decode(streamData)["data"] as Map;
+      for (var quality in qulityList) {
+        List<String> urls = [];
+        var flvUrl =
+            qualityData[quality["sdk_key"]]?["main"]?["flv"]?.toString();
+
+        if (flvUrl != null && flvUrl.isNotEmpty) {
+          urls.add(flvUrl);
+        }
+        var hlsUrl =
+            qualityData[quality["sdk_key"]]?["main"]?["hls"]?.toString();
+        if (hlsUrl != null && hlsUrl.isNotEmpty) {
+          urls.add(hlsUrl);
+        }
+        var qualityItem = LivePlayQuality(
+          quality: quality["name"],
+          sort: quality["level"],
+          data: urls,
+        );
+        if (urls.isNotEmpty) {
+          qualities.add(qualityItem);
+        }
+      }
     }
+    // var qualityData = json.decode(
+    //     detail.data["live_core_sdk_data"]["pull_data"]["stream_data"])["data"];
+
     qualities.sort((a, b) => b.sort.compareTo(a.sort));
     return qualities;
   }
@@ -327,33 +562,59 @@ class DouyinSite implements LiveSite {
       "browser_language": "zh-CN",
       "browser_platform": "Win32",
       "browser_name": "Edge",
-      "browser_version": "114.0.1823.58",
+      "browser_version": "125.0.0.0",
       "browser_online": "true",
       "engine_name": "Blink",
-      "engine_version": "114.0.0.0",
+      "engine_version": "125.0.0.0",
       "os_name": "Windows",
       "os_version": "10",
       "cpu_core_num": "12",
       "device_memory": "8",
       "platform": "PC",
-      "downlink": "4.7",
+      "downlink": "10",
       "effective_type": "4g",
       "round_trip_time": "100",
-      "webid": "7247041636524377637",
+      "webid": "7382872326016435738",
     });
-    var requlestUrl = await signUrl(uri.toString());
+    //var requlestUrl = await getAbogusUrl(uri.toString());
+    var requlestUrl = uri.toString();
+    var headResp = await HttpClient.instance
+        .head('https://live.douyin.com', header: headers);
+    var dyCookie = "";
+    headResp.headers["set-cookie"]?.forEach((element) {
+      var cookie = element.split(";")[0];
+      if (cookie.contains("ttwid")) {
+        dyCookie += "$cookie;";
+      }
+      if (cookie.contains("__ac_nonce")) {
+        dyCookie += "$cookie;";
+      }
+    });
+
     var result = await HttpClient.instance.getJson(
       requlestUrl,
       queryParameters: {},
       header: {
-        "Accept":
-            "text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.7",
-        "Authority": kDefaultAuthority,
-        "Referer": kDefaultReferer,
-        "Cookie": "__ac_nonce=${generateRandomString(21)}",
-        "User-Agent": kDefaultUserAgent,
+        "Authority": 'www.douyin.com',
+        'accept': 'application/json, text/plain, */*',
+        'accept-language': 'zh-CN,zh;q=0.9,en;q=0.8',
+        'cookie': dyCookie,
+        'priority': 'u=1, i',
+        'referer':
+            'https://www.douyin.com/search/${Uri.encodeComponent(keyword)}?type=live',
+        'sec-ch-ua':
+            '"Microsoft Edge";v="125", "Chromium";v="125", "Not.A/Brand";v="24"',
+        'sec-ch-ua-mobile': '?0',
+        'sec-ch-ua-platform': '"Windows"',
+        'sec-fetch-dest': 'empty',
+        'sec-fetch-mode': 'cors',
+        'sec-fetch-site': 'same-origin',
+        'user-agent': kDefaultUserAgent,
       },
     );
+    if (result == "" || result == 'blocked') {
+      throw Exception("抖音直播搜索被限制，请稍后再试");
+    }
     var items = <LiveRoomItem>[];
     for (var item in result["data"] ?? []) {
       var itemData = json.decode(item["lives"]["rawdata"].toString());
@@ -398,18 +659,32 @@ class DouyinSite implements LiveSite {
     return stringBuffer.toString();
   }
 
-  Future<String> signUrl(String url) async {
+  // 生成随机的数字
+  int generateRandomNumber(int length) {
+    var random = Random.secure();
+    var values = List<int>.generate(length, (i) => random.nextInt(10));
+    StringBuffer stringBuffer = StringBuffer();
+    for (var item in values) {
+      stringBuffer.write(item);
+    }
+    return int.tryParse(stringBuffer.toString()) ??
+        Random().nextInt(1000000000);
+  }
+
+  /// 读取A-Bogus签名后的URL
+  /// - [url] 原始URL
+  /// - 返回签名后的URL
+  ///
+  /// 服务端代码：https://github.com/dengmin/a-bogus，请自行部署后使用
+  Future<String> getAbogusUrl(String url) async {
     try {
-      // 发起一个签名请求
-      // 服务端代码：https://github.com/5ime/Tiktok_Signature
       var signResult = await HttpClient.instance.postJson(
-        "https://tk.nsapps.cn/",
+        "https://dy.nsapps.cn/abogus",
         queryParameters: {},
         header: {"Content-Type": "application/json"},
         data: {"url": url, "userAgent": kDefaultUserAgent},
       );
-      var requlestUrl = signResult["data"]["url"].toString();
-      return requlestUrl;
+      return signResult["data"]["url"];
     } catch (e) {
       CoreLog.error(e);
       return url;
