@@ -1,7 +1,6 @@
 import 'dart:convert';
 import 'dart:math';
 
-import 'package:simple_live_core/src/common/convert_helper.dart';
 import 'package:simple_live_core/src/common/http_client.dart';
 import 'package:simple_live_core/src/danmaku/huya_danmaku.dart';
 import 'package:simple_live_core/src/interface/live_danmaku.dart';
@@ -15,10 +14,14 @@ import 'package:simple_live_core/src/model/live_room_detail.dart';
 import 'package:simple_live_core/src/model/live_play_quality.dart';
 import 'package:simple_live_core/src/model/live_category_result.dart';
 import 'package:crypto/crypto.dart';
+import 'package:simple_live_core/src/model/tars/get_cdn_token_req.dart';
+import 'package:simple_live_core/src/model/tars/get_cdn_token_resp.dart';
+import 'package:tars_dart/tars/net/base_tars_http.dart';
 
 class HuyaSite implements LiveSite {
   final String kUserAgent =
       "Mozilla/5.0 (Linux; Android 11; Pixel 5) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/90.0.4430.91 Mobile Safari/537.36 Edg/117.0.0.0";
+  final BaseTarsHttp tupClient = BaseTarsHttp("http://wup.huya.com", "liveui");
 
   @override
   String id = "huya";
@@ -55,7 +58,18 @@ class HuyaSite implements LiveSite {
 
     List<LiveSubCategory> subs = [];
     for (var item in result["data"]) {
-      var gid = (asT<double?>(item["gid"])?.toInt() ?? 0).toString();
+      var gid = "";
+
+      if (item["gid"] is Map) {
+        gid = item["gid"]["value"].toString().split(",").first;
+      } else if (item["gid"] is double) {
+        gid = item["gid"].toInt().toString();
+      } else if (item["gid"] is int) {
+        gid = item["gid"].toString();
+      } else {
+        gid = item["gid"].toString();
+      }
+
       var subCategory = LiveSubCategory(
         id: gid,
         name: item["gameFullName"].toString(),
@@ -131,32 +145,36 @@ class HuyaSite implements LiveSite {
     //var url = getRealUrl(urlData.url);
 
     for (var item in urlData.bitRates) {
-      var urls = <String>[];
-      for (var line in urlData.lines) {
-        var src = line.line;
-        src += "/${line.streamName}";
-        if (line.lineType == HuyaLineType.flv) {
-          //src = src.replaceAll(".m3u8", ".flv");
-          src += ".flv";
-        }
-        if (line.lineType == HuyaLineType.hls) {
-          src += ".m3u8";
-        }
-        var parms = processAnticode(
-          line.lineType == HuyaLineType.flv
-              ? line.flvAntiCode
-              : line.hlsAntiCode,
-          urlData.uid,
-          line.streamName,
-        );
-        src += "?$parms";
-        if (item.bitRate > 0) {
-          src += "&ratio=${item.bitRate}";
-        }
-        urls.add(src);
-      }
+      // var urls = <String>[];
+      // for (var line in urlData.lines) {
+      //   var src = line.line;
+      //   src += "/${line.streamName}";
+      //   if (line.lineType == HuyaLineType.flv) {
+      //     //src = src.replaceAll(".m3u8", ".flv");
+      //     src += ".flv";
+      //   }
+      //   if (line.lineType == HuyaLineType.hls) {
+      //     src += ".m3u8";
+      //   }
+      //   var parms = processAnticode(
+      //     line.lineType == HuyaLineType.flv
+      //         ? line.flvAntiCode
+      //         : line.hlsAntiCode,
+      //     urlData.uid,
+      //     line.streamName,
+      //   );
+      //   src += "?$parms";
+      //   if (item.bitRate > 0) {
+      //     src += "&ratio=${item.bitRate}";
+      //   }
+      //   urls.add(src);
+      // }
+
       qualities.add(LivePlayQuality(
-        data: urls,
+        data: {
+          "urls": urlData.lines,
+          "bitRate": item.bitRate,
+        },
         quality: item.name,
       ));
     }
@@ -168,7 +186,27 @@ class HuyaSite implements LiveSite {
   Future<List<String>> getPlayUrls(
       {required LiveRoomDetail detail,
       required LivePlayQuality quality}) async {
-    return quality.data as List<String>;
+    var ls = <String>[];
+    for (var element in quality.data["urls"]) {
+      var line = element as HuyaLineModel;
+      var url = await getPlayUrl(line, quality.data["bitRate"]);
+      ls.add(url);
+    }
+    return ls;
+  }
+
+  Future<String> getPlayUrl(HuyaLineModel line, int bitRate) async {
+    var req = GetCdnTokenReq();
+    req.cdnType = line.cdnType;
+    req.streamName = line.streamName;
+    var resp =
+        await tupClient.tupRequest("getCdnTokenInfo", req, GetCdnTokenResp());
+    var url =
+        '${line.line}/${resp.streamName}.flv?${resp.flvAntiCode}&codec=264';
+    if (bitRate > 0) {
+      url += "&ratio=$bitRate";
+    }
+    return url;
   }
 
   @override
@@ -209,28 +247,18 @@ class HuyaSite implements LiveSite {
 
   @override
   Future<LiveRoomDetail> getRoomDetail({required String roomId}) async {
-    var resultText = await HttpClient.instance.getText(
-      "https://m.huya.com/$roomId",
-      queryParameters: {},
-      header: {
-        "user-agent": kUserAgent,
-      },
-    );
-    var text = RegExp(r"window\.HNF_GLOBAL_INIT.=.\{(.*?)\}.</script>",
-            multiLine: false)
-        .firstMatch(resultText)
-        ?.group(1);
-    var jsonObj = json.decode("{$text}");
-    var title =
-        jsonObj["roomInfo"]["tLiveInfo"]["sIntroduction"]?.toString() ?? "";
+    var roomInfo = await _getRoomInfo(roomId);
+    var tLiveInfo = roomInfo["roomInfo"]["tLiveInfo"];
+    var tProfileInfo = roomInfo["roomInfo"]["tProfileInfo"];
+
+    var title = tLiveInfo["sIntroduction"]?.toString() ?? "";
     if (title.isEmpty) {
-      title = jsonObj["roomInfo"]["tLiveInfo"]["sRoomName"]?.toString() ?? "";
+      title = tLiveInfo["sRoomName"]?.toString() ?? "";
     }
     var huyaLines = <HuyaLineModel>[];
     var huyaBiterates = <HuyaBitRateModel>[];
     //读取可用线路
-    var lines = jsonObj["roomInfo"]["tLiveInfo"]["tLiveStreamInfo"]
-        ["vStreamInfo"]["value"];
+    var lines = tLiveInfo["tLiveStreamInfo"]["vStreamInfo"]["value"];
     for (var item in lines) {
       if ((item["sFlvUrl"]?.toString() ?? "").isNotEmpty) {
         huyaLines.add(HuyaLineModel(
@@ -239,13 +267,13 @@ class HuyaSite implements LiveSite {
           flvAntiCode: item["sFlvAntiCode"].toString(),
           hlsAntiCode: item["sHlsAntiCode"].toString(),
           streamName: item["sStreamName"].toString(),
+          cdnType: item["sCdnType"].toString(),
         ));
       }
     }
 
     //清晰度
-    var biterates = jsonObj["roomInfo"]["tLiveInfo"]["tLiveStreamInfo"]
-        ["vBitRateInfo"]["value"];
+    var biterates = tLiveInfo["tLiveStreamInfo"]["vBitRateInfo"]["value"];
     for (var item in biterates) {
       var name = item["sDisplayName"].toString();
       if (name.contains("HDR")) {
@@ -257,38 +285,65 @@ class HuyaSite implements LiveSite {
       ));
     }
 
+    var topSid = roomInfo["topSid"];
+    var subSid = roomInfo["subSid"];
+
+    return LiveRoomDetail(
+      cover: tLiveInfo["sScreenshot"].toString(),
+      online: tLiveInfo["lTotalCount"],
+      roomId: tLiveInfo["lProfileRoom"].toString(),
+      title: title,
+      userName: tProfileInfo["sNick"].toString(),
+      userAvatar: tProfileInfo["sAvatar180"].toString(),
+      introduction: tLiveInfo["sIntroduction"].toString(),
+      notice: roomInfo["welcomeText"].toString(),
+      status: roomInfo["roomInfo"]["eLiveStatus"] == 2,
+      data: HuyaUrlDataModel(
+        url:
+            "https:${utf8.decode(base64.decode(roomInfo["roomProfile"]["liveLineUrl"].toString()))}",
+        lines: huyaLines,
+        bitRates: huyaBiterates,
+        uid: getUid(t: 13, e: 10),
+      ),
+      danmakuData: HuyaDanmakuArgs(
+        ayyuid: tLiveInfo["lYyid"] ?? 0,
+        topSid: topSid ?? 0,
+        subSid: subSid ?? 0,
+      ),
+      url: "https://www.huya.com/$roomId",
+    );
+  }
+
+  Future<Map> _getRoomInfo(String roomId) async {
+    var resultText = await HttpClient.instance.getText(
+      "https://m.huya.com/$roomId",
+      queryParameters: {},
+      header: {
+        "user-agent": kUserAgent,
+      },
+    );
+    var text = RegExp(
+            r"window\.HNF_GLOBAL_INIT.=.\{[\s\S]*?\}[\s\S]*?</script>",
+            multiLine: false)
+        .firstMatch(resultText)
+        ?.group(0);
+    var jsonText = text!
+        .replaceAll(RegExp(r"window\.HNF_GLOBAL_INIT.=."), '')
+        .replaceAll("</script>", "")
+        .replaceAllMapped(RegExp(r'function.*?\(.*?\).\{[\s\S]*?\}'), (match) {
+      return '""';
+    });
+
+    var jsonObj = json.decode(jsonText);
     var topSid = int.tryParse(
         RegExp(r'lChannelId":([0-9]+)').firstMatch(resultText)?.group(1) ??
             "0");
     var subSid = int.tryParse(
         RegExp(r'lSubChannelId":([0-9]+)').firstMatch(resultText)?.group(1) ??
             "0");
-
-    return LiveRoomDetail(
-        cover: jsonObj["roomInfo"]["tLiveInfo"]["sScreenshot"].toString(),
-        online: jsonObj["roomInfo"]["tLiveInfo"]["lTotalCount"],
-        roomId: jsonObj["roomInfo"]["tLiveInfo"]["lProfileRoom"].toString(),
-        title: title,
-        userName: jsonObj["roomInfo"]["tProfileInfo"]["sNick"].toString(),
-        userAvatar:
-            jsonObj["roomInfo"]["tProfileInfo"]["sAvatar180"].toString(),
-        introduction:
-            jsonObj["roomInfo"]["tLiveInfo"]["sIntroduction"].toString(),
-        notice: jsonObj["welcomeText"].toString(),
-        status: jsonObj["roomInfo"]["eLiveStatus"] == 2,
-        data: HuyaUrlDataModel(
-          url:
-              "https:${utf8.decode(base64.decode(jsonObj["roomProfile"]["liveLineUrl"].toString()))}",
-          lines: huyaLines,
-          bitRates: huyaBiterates,
-          uid: getUid(t: 13, e: 10),
-        ),
-        danmakuData: HuyaDanmakuArgs(
-          ayyuid: jsonObj["roomInfo"]["tLiveInfo"]["lYyid"] ?? 0,
-          topSid: topSid ?? 0,
-          subSid: subSid ?? 0,
-        ),
-        url: "https://www.huya.com/$roomId");
+    jsonObj["topSid"] = topSid;
+    jsonObj["subSid"] = subSid;
+    return jsonObj;
   }
 
   @override
@@ -368,16 +423,8 @@ class HuyaSite implements LiveSite {
 
   @override
   Future<bool> getLiveStatus({required String roomId}) async {
-    var resultText = await HttpClient.instance
-        .getText("https://m.huya.com/$roomId", queryParameters: {}, header: {
-      "user-agent": kUserAgent,
-    });
-    var text = RegExp(r"window\.HNF_GLOBAL_INIT.=.\{(.*?)\}.</script>",
-            multiLine: false)
-        .firstMatch(resultText)
-        ?.group(1);
-    var jsonObj = json.decode("{$text}");
-    return jsonObj["roomInfo"]["eLiveStatus"] == 2;
+    var roomInfo = await _getRoomInfo(roomId);
+    return roomInfo["roomInfo"]["eLiveStatus"] == 2;
   }
 
   /// 匿名登录获取uid
@@ -465,8 +512,8 @@ class HuyaSite implements LiveSite {
     // 通过ChatGPT转换的Dart代码
     var query = Uri.splitQueryString(anticode);
 
-    query["t"] = "100";
-    query["ctype"] = "huya_live";
+    query["t"] = "103";
+    query["ctype"] = "tars_mobile";
 
     final wsTime = (DateTime.now().millisecondsSinceEpoch ~/ 1000 + 21600)
         .toRadixString(16);
@@ -490,14 +537,18 @@ class HuyaSite implements LiveSite {
       "ctype": query["ctype"]!,
       "ver": "1",
       "fs": query["fs"]!,
-      "sphdcdn": query["sphdcdn"] ?? "",
-      "sphdDC": query["sphdDC"] ?? "",
-      "sphd": query["sphd"] ?? "",
-      "exsphd": query["exsphd"] ?? "",
+      // "sphdcdn": query["sphdcdn"] ?? "",
+      // "sphdDC": query["sphdDC"] ?? "",
+      // "sphd": query["sphd"] ?? "",
+      // "exsphd": query["exsphd"] ?? "",
+      "dMod": "mseh-0",
+      "sdkPcdn": "1_1",
       "uid": uid,
       "uuid": getUUid(),
       "t": query["t"]!,
-      "sv": "2110211124"
+      "sv": "202411221719",
+      "sdk_sid": "1732862566708",
+      "a_block": "0"
     }).query;
   }
 
@@ -529,10 +580,12 @@ enum HuyaLineType {
 
 class HuyaLineModel {
   final String line;
+  final String cdnType;
   final String flvAntiCode;
   final String hlsAntiCode;
   final String streamName;
   final HuyaLineType lineType;
+  int bitRate;
 
   HuyaLineModel({
     required this.line,
@@ -540,6 +593,8 @@ class HuyaLineModel {
     required this.flvAntiCode,
     required this.hlsAntiCode,
     required this.streamName,
+    required this.cdnType,
+    this.bitRate = 0,
   });
 }
 
